unitest build: Add separate build/run targets in most.Makefile
Before change, the only option is to build and run all unitests, or
build and run a test suite, or build and run a single test.

We would like to have an option to have a separate build target and run target.
This will be useful while trying to attach to IDEs and debuggers. Some developers
would like only to build a single unitest without running it, and then attach to it.

* Added build all tests or run all tests targets to build or run all tests separately.
* Added build test suite or run test suite targets to build or run tests suits separately.
* Added build single test or run single test